Add columns to Pot to support split pots
For the first time, I needed to make some manual changes to the
migration file created by Flask-Migrate. Alembic (used by Flask-Migrate)
cannot create enum data types as column additions, only on new
tables. I needed to write some custom code to create the column
type. This is a significant point of failure if I ever change the
values in this enum.

I also needed to delete all data in my dev database for the
migrations to run correctly. Not a problem now, but it's a
deal breaker in production.

https://trello.com/c/MKiny9KZ
Tests FAILING