Ensure hand.betting_rounds ordered by id
Was having a problem where some hands would never end. The problem was that the
code to check if creating a new betting round was allowed just looked
for the last item in hand.betting_rounds, which I guess was not guaranteed
to be the most recent one created. Explicitly ordering ascending by ID
does the trick. I don't know why sometimes this list got out
of order and not other times, but it doesn't matter.

https://trello.com/c/rtAb01eC