Fix  at max slider position error
The problem was that I was always representing one too many bets on the slider. If
I called panel.setBets with a list of five bets, there were six total positions the
slider could take, and the max position was always reporting an index of
bets.length + 1. Need to account for difference between length and max index.