Rethink how emulator works
I was jumping through a bunch of hoops making sure the panel was
always up to date when the emulator is enabled, even though the
game explicitly didn't update the panel except for when the
user player was up next when the emulator was disabled. The whole
thing was just heavier than it had to be.

My solution was to always keep the panel up to date, regardless
of emulator state. The overhead should be minimal. The main benefit
here is that it makes emulation really simple -- all we need to
do is disable panel hiding.