Create JWT on show_table not join_table
Previously, I was creating JWTs as players joined the table,
which meant that the joining itself was almost completely
unguarded. This was likely just a temporary allowance for dev
purposes, but still. It also made the assumption that only
Actions needed to be authenticated, which isn't true. Buy-ins
and cash-outs need to be authenticated, as will chat and settings
changes eventually.

So now we create the JWT when the user first shows the table. Any
user who is viewing a table has first been logged in and then
verified to be part of the group that owns the table, so they
are always allowed to attempt to join the table.

This change meant that I had to rework how we go from a token to
a Player object, since we pass through the User object now. Some
more defensive measures were required to make sure that a user
who is currently sitting at another table owned by the group can't
take actions on this table.