Add ChipManager option to keep chips static
It's okay that the chips played in front of the player change with each action.
It's like the player is updating their bet as raises and calls happen. But once
chips are in the pot, they should be static instead of updating every time the
pot increases. Five whites should not be swapped for one red. This fixes that.