Fix broken migration
New learning: It's unsafe to import models directly into migrations
for the purposes of inserting data, because although migrations
represent a snapshot in time of when the migration was created,
the codebase has likely progressed since then. In this case
specifically, I was querying the  table to insert a
default  (as it's a non-null field). This migration
worked when I orginally wrote it. But the very next migration
added a change to the  table, which is referenced
by , adding a new column to the former. Importing the up-
to-date model meant that the code expected this new column to be
present, but it was not present at that point in the migration
history. Problem.

My solution was to avoid over-engineering by hardcoding the
specific change I wanted to make. Instead of using SQLAlchemy
to query the tables represented by the models, I just wrote
raw SQL to update the rows I wanted to update. Brittle, maybe,
but less fragile than what I wrote the first time.