Allow player update without chip changes
I was having problems stemming from clearing chips too soon. The takeChips
animation depends on each player having a stack of chips in front of them.
The animation fires on a delay, so even though the chips existed at the
time the code ran, by the time the animation fired they'd been cleared.