Choreograph transition from gatherChips animation to pay player animation
I originally was creating a class called EventRegister which
was going to be a way to globally record the existence of in-progress
animations so that other parts of the system could react to their
completion. For some reason, it wasn't working, so I changed
courses. Instead, I just added a handComplete boolean to the
roundComplete event and if it's present, I do not trigger
the gatherChips animation there. Rather, I wait until the
handComplete event, so I can more closely control how the two
animations fire in relation to each other.