Animate the reveal of cards on the board
Previously, board cards just appeared when they were dealt and
it was a bit jarring. This commit makes them animate in. It
looks pretty good. I'm having a ton of timing problems though,
so I'm stashing this on a branch and going back to master to try
to sort them out.