Fix signal dispatch timing problems
I was having trouble with the showdown choreography because the
signal which fires to tell listeners that the chips have been
collected into the pot was being fired before any listeners had
been added. Creating a class which automatically executes listeners
if added to a previously-dispatched signal fixed the problem.