Enable posting blinds via panel
Now there are two states the panel can be in, one for blinds and one
for betting. Blind amounts are no longer specified by the client, but
instead are set dynamically by the back end. Client just sends
blind request and the back end figures it out.

This enables me to remove two of the temporary buttons on the left.